Fix: expand/collapse state for CRD sidebar items (#1751)
* Moving SidebarNavItem component to its own file


* Using id prop to preserve expanding state
